take palantir's feedbacks in account + fix some bugs
