+++ com.palantir.typescript/src/com/palantir/typescript/preferences/ProjectPreferenceStore.java
@@ -17,21 +17,11 @@
 package com.palantir.typescript.preferences;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.palantir.typescript.TypeScriptPlugin.logError;
-import static com.palantir.typescript.TypeScriptPlugin.logInfo;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.io.OutputStream;
-import java.security.MessageDigest;
-import java.util.Collection;
-import java.util.Map;
-import java.util.Objects;
 
-import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IProject;
-import org.eclipse.core.resources.IResource;
 import org.eclipse.core.resources.ProjectScope;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.core.runtime.preferences.IScopeContext;
@@ -39,16 +29,8 @@ import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.preference.PreferenceStore;
 import org.osgi.service.prefs.BackingStoreException;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Joiner;
-import com.google.common.collect.BiMap;
-import com.google.common.collect.HashBiMap;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.io.CharStreams;
 import com.palantir.typescript.IPreferenceConstants;
 import com.palantir.typescript.TypeScriptPlugin;
-import com.palantir.typescript.TypeScriptProjects;
 
 /**
  * An adapter for {@link PreferenceStore} to allow a preference page to be used as a property page.
@@ -65,21 +47,9 @@ import com.palantir.typescript.TypeScriptProjects;
  */
 public final class ProjectPreferenceStore extends PreferenceStore {
 
-    private static final BiMap<String, String> PREFERENCE_TO_TSCONFIG_PATH = createPreferenceToTsConfigPathMap();
-
-    private static final Map<String, String> TSCONFIG_PATH_TO_PREFERENCE = PREFERENCE_TO_TSCONFIG_PATH.inverse();
-
-    private static final Map<String, String> TSCONFIG_TARGET_MAPPING = ImmutableMap.of(
-        "ES3", "ECMASCRIPT3", "ES5", "ECMASCRIPT5", "ES6", "ECMASCRIPT6", "ES2015", "ECMASCRIPT6");
-
-    private static final Map<String, String> TSCONFIG_MODULE_KIND_MAPPING = ImmutableMap.of(
-        "COMMONJS", "COMMONSJS", "ES2015", "ES6");
-
-    private static final Map<String, String> TSCONFIG_MODULE_RESOLUTION_MAPPING = ImmutableMap.of(
-        "NODEJS", "NODE_JS");
-
     private final IProject project;
     private final IPreferenceStore preferenceStore;
+    private final TsConfigPreferences tsConfigPreferences;
 
     private boolean projectSpecificSettings;
 
@@ -96,6 +66,7 @@ public final class ProjectPreferenceStore extends PreferenceStore {
 
         this.project = project;
         this.preferenceStore = preferenceStore;
+        this.tsConfigPreferences = new TsConfigPreferences(project);
 
         IEclipsePreferences projectPreferences = this.getProjectPreferences();
         this.projectSpecificSettings = (projectPreferences.get(sentinelPropertyName, null) != null);
@@ -227,10 +198,13 @@ public final class ProjectPreferenceStore extends PreferenceStore {
             return;
         }
 
-        boolean useTsConfigFile = isUsingTsConfigFile();
-        if (useTsConfigFile && isTsConfigPreference(name)) {
-            if (didTsConfigFileChanged()) {
-                reloadTsConfigFile();
+        if (isUsingTsConfigFile() && tsConfigPreferences.isTsConfigPreference(name)) {
+            IEclipsePreferences projectPreferences = this.getProjectPreferences();
+            String value = tsConfigPreferences.getValue(name);
+            if (value == null) {
+                projectPreferences.remove(name);
+            } else {
+                projectPreferences.put(name, value);
             }
         }
 
@@ -255,199 +229,8 @@ public final class ProjectPreferenceStore extends PreferenceStore {
         }
     }
 
-    public boolean reloadTsConfigFile() {
-        logInfo("reload tsconfig file");
-
-        IFile tsConfigFile = null;
-        InputStream tsConfigStream = null;
-        boolean loaded = false;
-        try {
-            tsConfigFile = getTsConfigFile();
-            if (tsConfigFile.exists()) {
-
-                if (!tsConfigFile.isSynchronized(IResource.DEPTH_ZERO)) {
-                    tsConfigFile.refreshLocal(IResource.DEPTH_ZERO, null);
-                }
-
-                // refresh tsconfig cache infos
-                IEclipsePreferences projectPreferences = this.getProjectPreferences();
-                projectPreferences.put(IPreferenceConstants.PREFERENCE_STORE_TS_CONFIG_HASH, getFileSHA1(tsConfigFile));
-                projectPreferences.putLong(IPreferenceConstants.PREFERENCE_STORE_TS_CONFIG_LAST_MODIFICATION_TIME,
-                    tsConfigFile.getModificationStamp());
-
-                // read tsconfig JSON content
-                tsConfigStream = tsConfigFile.getContents();
-                String tsConfigContent = CharStreams.toString(new InputStreamReader(tsConfigStream, tsConfigFile.getCharset()));
-
-                // convert tsconfig JSON to Map
-                ObjectMapper mapper = new ObjectMapper();
-                Map<String, Object> tsConfigEntries = mapper.readValue(tsConfigContent,
-                    new TypeReference<Map<String, Object>>() {
-                    });
-
-                // resets preferences to default
-                resetTsConfigDefaultPreferences();
-
-                // reads preferences from tsconfig
-                decodeTsConfigEntries("", tsConfigEntries);
-
-                loaded = true;
-            }
-
-        } catch (Exception e) {
-            logError("Cannot reload ts config file '" + tsConfigFile + "'", e);
-        } finally {
-            if (tsConfigStream != null) {
-                try {
-                    tsConfigStream.close();
-                } catch (IOException e) {
-                    System.err.println("error while releasing tsconfig stream");
-                }
-            }
-        }
-
-        return loaded;
-    }
-
-    private void resetTsConfigDefaultPreferences() {
-        IEclipsePreferences projectPreferences = this.getProjectPreferences();
-        for (String preferenceKey : PREFERENCE_TO_TSCONFIG_PATH.keySet()) {
-            String defaultValue = this.getDefaultString(preferenceKey);
-            this.setValue(preferenceKey, defaultValue);
-            projectPreferences.put(preferenceKey, defaultValue);
-        }
-        System.out.println("tsconfig preferences reset: " + this);
-    }
-
-    private void decodeTsConfigEntries(String jsonTreePath, Map<String, Object> entries) {
-
-        IEclipsePreferences projectPreferences = this.getProjectPreferences();
-
-        for (Map.Entry<String, Object> tsConfigEntry : entries.entrySet()) {
-            if (tsConfigEntry.getValue() instanceof Map) {
-                decodeTsConfigEntries(jsonTreePath + tsConfigEntry.getKey() + ".", (Map) tsConfigEntry.getValue());
-            } else if (isSupportedTsConfigPath(jsonTreePath + tsConfigEntry.getKey())) {
-
-                String matchingPreference = TSCONFIG_PATH_TO_PREFERENCE.get(jsonTreePath + tsConfigEntry.getKey());
-
-                String value = tsConfigValueToPreferenceValue(tsConfigEntry.getValue(), matchingPreference);
-                logInfo("setting preference " + matchingPreference + " to " + value);
-                this.setValue(matchingPreference, value);
-                projectPreferences.put(matchingPreference, value);
-            }
-        }
-    }
-
-    public String tsConfigValueToPreferenceValue(Object tsConfigValue, String matchingPreference) {
-        String value = null;
-        if (tsConfigValue != null) {
-
-            if (tsConfigValue instanceof Collection) {
-                value = Joiner.on(TypeScriptProjects.BUILD_PATH_SPEC_SEPARATOR).join((Collection) tsConfigValue);
-            } else {
-                value = tsConfigValue.toString();
-            }
-
-            // TODO : awful mapping, we should rename our enums to match standard options / tsconfig values
-            if (matchingPreference.equals(IPreferenceConstants.COMPILER_TARGET)) {
-                value = value.toUpperCase();
-                if (TSCONFIG_TARGET_MAPPING.containsKey(value)) {
-                    value = TSCONFIG_TARGET_MAPPING.get(value);
-                }
-            } else if (matchingPreference.equals(IPreferenceConstants.COMPILER_MODULE)) {
-                value = value.toUpperCase();
-                if (TSCONFIG_MODULE_KIND_MAPPING.containsKey(value)) {
-                    value = TSCONFIG_MODULE_KIND_MAPPING.get(value);
-                }
-            } else if (matchingPreference.equals(IPreferenceConstants.COMPILER_MODULE_RESOLUTION)) {
-                value = value.toUpperCase();
-                if (TSCONFIG_MODULE_RESOLUTION_MAPPING.containsKey(value)) {
-                    value = TSCONFIG_MODULE_RESOLUTION_MAPPING.get(value);
-                }
-            } else if (matchingPreference.equals(IPreferenceConstants.COMPILER_JSX)) {
-                value = value.toUpperCase();
-            }
-        }
-        return value;
-    }
-
-    private boolean isTsConfigPreference(String name) {
-        return PREFERENCE_TO_TSCONFIG_PATH.containsKey(name);
-    }
-
-    private boolean isSupportedTsConfigPath(String tsConfigJsonPath) {
-        return TSCONFIG_PATH_TO_PREFERENCE.containsKey(tsConfigJsonPath);
-    }
-
-    private String getFileSHA1(IFile file) {
-
-        if (!file.exists()) {
-            return null;
-        }
-
-        InputStream fileInputStream = null;
-        try {
-            fileInputStream = file.getContents();
-            MessageDigest md = MessageDigest.getInstance("SHA1");
-            byte[] dataBytes = new byte[1024];
-
-            int nread = 0;
-
-            while ((nread = fileInputStream.read(dataBytes)) != -1) {
-                md.update(dataBytes, 0, nread);
-            }
-
-            byte[] mdbytes = md.digest();
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < mdbytes.length; i++) {
-                sb.append(Integer.toString((mdbytes[i] & 0xff) + 0x100, 16).substring(1));
-            }
-            return sb.toString();
-        } catch (Exception e) {
-            logError("Cannot sha1 '" + file + "'", e);
-            return null;
-        } finally {
-            try {
-                if (fileInputStream != null) {
-                    fileInputStream.close();
-                }
-            } catch (Exception e) {
-            }
-        }
-    }
-
-    private boolean didTsConfigFileChanged() {
-        IEclipsePreferences projectPreferences = this.getProjectPreferences();
-
-        IFile tsConfigFile = getTsConfigFile();
-        boolean changed = true;
-        if (tsConfigFile.exists()) {
-            changed = false;
-            try {
-                long previousLastModTimestamp = projectPreferences
-                    .getLong(IPreferenceConstants.PREFERENCE_STORE_TS_CONFIG_LAST_MODIFICATION_TIME, -1);
-                long currentLastModTimestamp = tsConfigFile.getModificationStamp();
-                if (previousLastModTimestamp != currentLastModTimestamp) {
-
-                    // optimization: get hash only if mod timestamp changed
-                    String previousHash = projectPreferences.get(IPreferenceConstants.PREFERENCE_STORE_TS_CONFIG_HASH, null);
-                    String currentHash = getFileSHA1(tsConfigFile);
-                    if (!Objects.equals(previousHash, currentHash)) {
-                        changed = true;
-                    }
-                }
-
-            } catch (Exception e) {
-                logError("Cannot tell if tsconfig file changed '" + tsConfigFile + "'", e);
-            }
-        }
-
-        return changed;
-    }
-
-    public IFile getTsConfigFile() {
-        IFile tsConfigFile = this.project.getFile("tsconfig.json");
-        return tsConfigFile;
+    public TsConfigPreferences getTsConfigPreferences() {
+        return tsConfigPreferences;
     }
 
     private void writeProperties() throws IOException {
@@ -474,38 +257,6 @@ public final class ProjectPreferenceStore extends PreferenceStore {
         IScopeContext projectScope = new ProjectScope(this.project);
 
         return projectScope.getNode(TypeScriptPlugin.ID);
-    }
-
-    private static BiMap<String, String> createPreferenceToTsConfigPathMap() {
-        BiMap<String, String> map = HashBiMap.<String, String> create();
-
-        map.put(IPreferenceConstants.BUILD_PATH_FILES, "files");
-        map.put(IPreferenceConstants.BUILD_PATH_INCLUDE, "include");
-        map.put(IPreferenceConstants.BUILD_PATH_EXCLUDE, "exclude");
-
-        map.put(IPreferenceConstants.COMPILER_COMPILE_ON_SAVE, "compileOnSave");
-        map.put(IPreferenceConstants.COMPILER_DECLARATION, "compilerOptions.declaration");
-        map.put(IPreferenceConstants.COMPILER_EXPERIMENTAL_DECORATORS, "compilerOptions.experimentalDecorators");
-        map.put(IPreferenceConstants.COMPILER_EMIT_DECORATOR_METADATA, "compilerOptions.emitDecoratorMetadata");
-        map.put(IPreferenceConstants.COMPILER_INLINE_SOURCE_MAP, "compilerOptions.inlineSourceMap");
-        map.put(IPreferenceConstants.COMPILER_INLINE_SOURCES, "compilerOptions.inlineSource");
-        map.put(IPreferenceConstants.COMPILER_JSX, "compilerOptions.jsx");
-        map.put(IPreferenceConstants.COMPILER_MODULE, "compilerOptions.module");
-        map.put(IPreferenceConstants.COMPILER_MODULE_RESOLUTION, "compilerOptions.moduleResolution");
-        map.put(IPreferenceConstants.COMPILER_NO_EMIT_ON_ERROR, "compilerOptions.noEmitOnError");
-        map.put(IPreferenceConstants.COMPILER_NO_FALLTHROUGH_CASES_IN_SWITCH, "compilerOptions.noFallthroughCasesInSwitch");
-        map.put(IPreferenceConstants.COMPILER_NO_IMPLICIT_ANY, "compilerOptions.noImplicitAny");
-        map.put(IPreferenceConstants.COMPILER_NO_IMPLICIT_RETURNS, "compilerOptions.noImplicitReturns");
-        map.put(IPreferenceConstants.COMPILER_NO_LIB, "compilerOptions.noLib");
-        map.put(IPreferenceConstants.COMPILER_OUT_DIR, "compilerOptions.outDir");
-        map.put(IPreferenceConstants.COMPILER_OUT_FILE, "compilerOptions.outFile");
-        map.put(IPreferenceConstants.COMPILER_REMOVE_COMMENTS, "compilerOptions.removeComments");
-        map.put(IPreferenceConstants.COMPILER_SOURCE_MAP, "compilerOptions.sourceMap");
-        map.put(IPreferenceConstants.COMPILER_SUPPRESS_EXCESS_PROPERTY_ERRORS, "compilerOptions.suppressExcessPropertyErrors");
-        map.put(IPreferenceConstants.COMPILER_SUPPRESS_IMPLICIT_ANY_INDEX_ERRORS, "compilerOptions.suppressImplicitAnyIndexErrors");
-        map.put(IPreferenceConstants.COMPILER_TARGET, "compilerOptions.target");
-
-        return map;
     }
 
     @Override
