+++ com.palantir.typescript/src/com/palantir/typescript/resources.properties
@@ -78,11 +78,11 @@ preferences.formatter.insert.space.after.opening.and.before.closing.template.str
 preferences.formatter.place.open.brace.on.new.line.for.functions = Place open brace on new line for functions
 preferences.formatter.place.open.brace.on.new.line.for.control.blocks = Place open brace on new line for control blocks
 
-preferences.tsconfig.loadError = An error occurred while loading tsconfig.json file. Does it exist?
-
 preferences.general.node.path = Node path:
 preferences.general.node.path.dialog.message = The Node.js path has changed. Eclipse must be rebooted for this change to take effect.
 preferences.general.node.path.dialog.title = Node.js Path Changed
+
+preferences.tsconfig.loadError = An error occurred while loading tsconfig.json file. Does it exist?
 
 search.result.match = ({0,number,integer} matches)
